DEV-2322 Fix update metadata
The payload of the update metadata call needs to be a record update object.
Only the sidecar was given as an XML payload in the body. Instead of passing
an XML in the body, change it to multipart/form-data and add the missing
parameters as other form-data together with the sidecar XML.