DEV-2322 Fix error handling
When there is a unsuccessful response, meaning in the 4xx-5xx range,
there used to be an HTTPError that was thrown. The MediaHaven Python library
abstracts this, and throws its own MediaHavenException. Catch that exception
instead.

In case of the search request, a MediaHavenException was caught instead of
RequestException. Expand, so they are both properly handled.