DEV-2386 Add failing test for FAILED events
Failed events with wrong or missing metadata cause the service to crash
when attempting to parse the metadata (while getting the media type).
This error is caught and an InvalidEventException is returned.  Instead,
we should avoid attempting to parse the failed event in the first place.
Two failed test events were added, of which the minimal one is probably
the realistic one.  The one with valid metadata seems artificial and
probably wouldn't actually occur.  Since both failure modes (checking
the event status and failing to parse the metadata) result in the same
error being raised, we check the error message itself to create the
failing test.